Retry the reading of the pos response
We need to check if the LRC of the reponse is the same as the one
included in the message.

Adds a method to calculate the LRC od a message.
Refactor reply_ack to responde ACK or NACK depending on the result of
the LRC check.
Retry the reading when reply_ack sends a NACK.
Add Test for LoadKeys method